feat: Add compatibility inventory and command templates for GitHub address management
- Introduced a new `compatibility-inventory.md` to document preserved public contracts and removed surfaces.
- Created command templates in `command_templates.py` for various GitHub address actions, including address, review, threads, and classification.
- Implemented stable ID generation functions in `ids.py` for consistent identification of payloads.
- Added comprehensive tests for control plane workflows, including handling of fix-all operations, validation requirements, and stale thread management.
- Developed unit tests for stable ID generation to ensure order independence and hash consistency.
