+++ src/gh_address_cr/commands/__init__.py
@@ -1,1 +1,1 @@
-# legacy handlers
+# Command handler package.
