+++ src/gh_address_cr/core/workflow.py
@@ -2,12 +2,12 @@ from __future__ import annotations
 
 import json
 import re
-import shlex
 from datetime import datetime
 from pathlib import Path
 from typing import Any
 from uuid import uuid4
 
+from gh_address_cr.core import command_templates
 from gh_address_cr import (
     MAX_PARALLEL_CLAIMS,
     PROTOCOL_VERSION,
@@ -669,8 +669,7 @@ def fast_fix_matching_threads(
             payload={"matched_count": len(matches), "files": sorted(normalized_file_set)},
         )
     if not stale_only and not normalized_homogeneous_reason:
-        files_str = ",".join(sorted(normalized_file_set))
-        batch_command = f"gh-address-cr agent next {repo} {pr_number} --batch --agent-id <agent_id> --files {shlex.quote(files_str)}"
+        batch_command = command_templates.batch_next(repo, str(pr_number), files=sorted(normalized_file_set))
         next_action = (
             f"Run `{batch_command}` to claim the matching GitHub review threads and write a "
             "BatchActionResponse skeleton, then fill per-thread summary/why entries and submit it. "
@@ -687,14 +686,13 @@ def fast_fix_matching_threads(
                 "files": sorted(normalized_file_set),
                 "commands": {
                     "batch_next": batch_command,
-                    "submit_batch": f"gh-address-cr agent submit-batch {repo} {pr_number} --input <batch-response.json>",
-                    "fix_all_input": f"gh-address-cr agent fix-all {repo} {pr_number} --input <batch-response.json>",
+                    "submit_batch": command_templates.submit_batch(repo, str(pr_number), input_path="<batch-response.json>"),
+                    "fix_all_input": command_templates.fix_all_input(repo, str(pr_number), input_path="<batch-response.json>"),
                 },
             },
         )
     if not stale_only and normalized_homogeneous_reason and not _has_homogeneous_thread_bodies(matches):
-        files_str = ",".join(sorted(normalized_file_set))
-        batch_command = f"gh-address-cr agent next {repo} {pr_number} --batch --agent-id <agent_id> --files {shlex.quote(files_str)}"
+        batch_command = command_templates.batch_next(repo, str(pr_number), files=sorted(normalized_file_set))
         next_action = (
             f"Run `{batch_command}` to claim the matching GitHub review threads and write a "
             "BatchActionResponse skeleton with per-thread summary/why entries. The matched threads have missing "
@@ -711,8 +709,8 @@ def fast_fix_matching_threads(
                 "files": sorted(normalized_file_set),
                 "commands": {
                     "batch_next": batch_command,
-                    "submit_batch": f"gh-address-cr agent submit-batch {repo} {pr_number} --input <batch-response.json>",
-                    "fix_all_input": f"gh-address-cr agent fix-all {repo} {pr_number} --input <batch-response.json>",
+                    "submit_batch": command_templates.submit_batch(repo, str(pr_number), input_path="<batch-response.json>"),
+                    "fix_all_input": command_templates.fix_all_input(repo, str(pr_number), input_path="<batch-response.json>"),
                 },
             },
         )
